+++ sdk/CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Ed25519Signature` type no longer requires validated public key bytes to construct;
 - `SelfControlledAliasOutput` and `SelfDepositNft` conditions;
+- Input selection for > max native tokens;
 
 ## 1.1.3 - 2023-12-07